budgetbook/requirements.txt
Django>=5.2,<5.3
django-htmx>=1.27,<1.28
python-dotenv>=1.1,<2
whitenoise>=6.7,<7
